fix bug

+++ src/controller/user.controller.ts
@@ -117,7 +117,7 @@ export const deleteUser = async (req: Request, res: Response, next: NextFunction
     await PosterModel.deleteMany({ user_id: user?._id })
 
     const filepath = `./public/profil/${user?.imgProfil}`
-    if (fs.existsSync(filepath) || filepath !== './public/profil/default-fotoprofil.png') {
+    if (fs.existsSync(filepath) && filepath !== './public/profil/default-fotoprofil.png') {
       fs.unlinkSync(filepath)
     }
